cleanup dracula pass
Summary:This is the redex machinery necessary to clean up the leftovers from Dracula transformations.

The logic, as explained by rightparen :

  Here is the diff:  D3049241  The types that this diff should make removable are:
  com.facebook.groups.feed.protocol.FetchGroupInformationGraphQLInterfaces.AdminAwareGroup,
  com.facebook.groups.feed.protocol.FetchGroupInformationGraphQLModels.AdminAwareGroupModel,
  and all nested interfaces/classes within those types.  The transform should remove any types/methods/fields
  marked with RedirectedTo.  The transform also needs to remove references to these types from the 'implements'
  clause of types marked with RedirectedFrom.

Michael and I discovered that in addition to this, there are a few compulsory assignments of static fields that are to be deleted. Those are useless, so we can just strip them. Thus, you'll see the bit of the diff that removes sputs.

The benefits of this transform scale up as Dracula rolls out. Dracula's effectively turned off right now though, so no gains out of the gate.

For the fb4a APK built against D3049241, we got:

  DRAC :) Deleted 2 classes
  DRAC :) Deleted 123 methods
  DRAC :) Deleted 19 fields
  DRAC :| Removed 2 'implements' references of deleted classes
  DRAC :| Removed 1 sputs that referenced deleted fields

Reviewed By: bertmaher

Differential Revision: D3095877

fb-gh-sync-id: 56bf3cf1832170e698aa2368518eb5d78c986ba0
fbshipit-source-id: 56bf3cf1832170e698aa2368518eb5d78c986ba0