collect stats
Summary:Have redex emit (as json) stats on the number of dex items that it output.

This is a stepping stone for:

1) Integrating redex stats into scuba
2) Autoposting stats into phabricator
3) Autocomputing/posting differentials into phabricator, e.g. given APK vX, and Redex vN and vN', what impact did diff vN' have over vN?

Maybe there's a better place/way to do this (e.g. as a pass, or just leverage what interdex has), but anyway, I wanted to at least get something in there so I could start tinkering on the Sandcastle/scuba side.

Reviewed By: bertmaher

Differential Revision: D3104073

fb-gh-sync-id: df29b5656117889267a5d1e90ca53876c32759f8
fbshipit-source-id: df29b5656117889267a5d1e90ca53876c32759f8