Clean up synth for external references
Summary:Synth seemed pretty good wrt external refrences however I just added few checks and asserts to make sure we are good in the face of external references.
I did not make some of the assert always_assert because we should only be using concrete so it's more of a double check test.

Reviewed By: bertmaher

Differential Revision: D3110696

fb-gh-sync-id: 754bbce3351e68d135ab6d67a6c360c078f6538b
fbshipit-source-id: 754bbce3351e68d135ab6d67a6c360c078f6538b