Some cleanup of config reading
Summary:We're using folly::dynamic in a really verbose way, which makes it kind
of hard to figure out what our configs are doing.  This simplifies it quite a
bit.  For now I'm avoiding doing too much error-handling except in cases where
stuff is truly optional; in those cases I'm just catching the
TypeError/out_of_range and swallowing it.

Reviewed By: newobj

Differential Revision: D3112258

fb-gh-sync-id: ba09f4694e441002a4a0018422d9f0bae40f6680
fbshipit-source-id: ba09f4694e441002a4a0018422d9f0bae40f6680