SimpleInline simple...
Summary:In the process of reviewing external references I ended up changing a bit around the inliner. I mean it's all the same but its checks against the type system are reviewed a bit.
I like those more and they should be what was the intent always (and the code maybe never).
Results are slightly different so in keeping the same identical semantic I changed it (go figure).
I like this one so whatever

Reviewed By: bertmaher

Differential Revision: D3130409

fb-gh-sync-id: 78922a440b92054bdf338efd71b7b6cb27e6d0b4
fbshipit-source-id: 78922a440b92054bdf338efd71b7b6cb27e6d0b4