Fix single impl
Summary:Apparently and not surprising I am an idiot. SingleImpl (should it be renamed SimpleImpl?) really wants refs and not defs so by moving them to defs I broke it.
Essentially it's tracking signatures and not method defs and so it wants to know about any signature that is used.

Reviewed By: onomatopoeia

Differential Revision: D3140097

fb-gh-sync-id: e73b95b1e4032f845c18fc6b5e43de34c59f7398
fbshipit-source-id: e73b95b1e4032f845c18fc6b5e43de34c59f7398