Extend ProGaurd parser to parse full FB4A config and extract library jars
Summary: The current ProGuard parser only parses the merged_fb4a_proguard.txt configuration file but there are around 50 files that make up the full configuration. This revision takes the full dump of the ProGuard configuration produced at build time and propagated this is Redex. It modifies the ProGuard parser to parse the full configuration and it also extracts the list of library jars from the full configuration which is combined with the command line specified library jar (if given). The ProGuard parser can now also accept an empty ProGuard config without crashing.

Reviewed By: dariorussi, bertmaher

Differential Revision: D3143195

fb-gh-sync-id: f2d0cc85008840fcedabc52e5c5a847f1e9e65b5
fbshipit-source-id: f2d0cc85008840fcedabc52e5c5a847f1e9e65b5