Don't let single impl optimize interfaces in the primary DEX that would bring classes not defined in primary
Summary:SingleImpl was letting classes in any DEX be substitute for interfaces defined in the primary DEX. However that is problematic if the class loader is not set up yet to understand secondary DEXes, causing verification or loading errors.
So don't optimize those cases.
We have about ~70 interfaces in fb4a in this bucket

Reviewed By: bertmaher

Differential Revision: D3147623

fb-gh-sync-id: e55f3c657c18eddb0d0f1a439cffa32268663ff0
fbshipit-source-id: e55f3c657c18eddb0d0f1a439cffa32268663ff0