add no inline annotation logic and define an overall no optimizations annotation
Summary:Hard to make a diff more disgusting than this but it has merits in its disgrace.
This is giving us an easy way out of optimizations for certain optimizations.
Essentially introduces a concept of do not optimize at all that when applied to a class translates to all memebers of the class.
All of that for certain optimizations only! The one that care to look for it basically.
For now those are the one we know can be problematic like inline and static relo.
It also makes the inliner support random annotation to stop inlining if desired.
The PgoFile class is also taking a clownish shape being much more that the PGO stuff it was originally intended for.
We need to clean all of this up really soon, but this is it for now.
This picks (and so does not optimize) few methods in static relo and few in the inliner. Those seem to be around classes that were delicate (i.e. Lcom/facebook/acra/ErrorReporter$Api9Utils;) so it seems good

Reviewed By: newobj

Differential Revision: D3185010

fb-gh-sync-id: b83519aa9507cde4da615103152ac46cb1417429
fbshipit-source-id: b83519aa9507cde4da615103152ac46cb1417429