rebind refs revisited again
Summary:Made the rebinder work better with external refrences and correct semantic.
We have 3 modes of rebinding now, all related to virtual invocation:
NORMAL does not touch virtual invocation, not even array clone
MEDIUM rebinds virtual to the first definition up the hierarchy
HIGH rebinds virtual to the top definition up the hierarchy (method introduction)
The number of refs saved is highest in HIGH.
There are a set of issues related to rebinding virtual which I am going to post about soon.

We are running with HIGH mode. Or at least that is what the config is set up to with this diff

Reviewed By: shanenay

Differential Revision: D3189892

fb-gh-sync-id: 0f8df4f27ac8a314a935ea8b4c755168cdb01c6e
fbshipit-source-id: 0f8df4f27ac8a314a935ea8b4c755168cdb01c6e