--redex-binary is for binary
Summary:OSS users are frequently confused when they try to do something like:
```
redex in.apk out.apk
```
And it tells them that "redex-all is not found".

When you think about it, it's pretty weird that the path to the binary is a
positional argument, especially since it's not usually needed.  Let's make it a
named argument instead.

Reviewed By: dcolascione

Differential Revision: D3200881

fb-gh-sync-id: 2c12cfd8240c038328e87e56bcff11a3d9c31445
fbshipit-source-id: 2c12cfd8240c038328e87e56bcff11a3d9c31445