use more conservative criteria to determine if methods collide
Summary:Make StaticRelo more conservative by requiring that methods also don't alias on line numbers. This is because on a stack trace we just get the line number back, not the full signature. So we can't disambiguate 2 methods which would show up as foo.a(bar.java:20) from each other, even if their signatures differ.

This cuts the number of moved methods by ~350 on fb4a right now.

Reviewed By: newobj

Differential Revision: D3205867

fb-gh-sync-id: bfaa5203fc7d998635c3c3f398cc180c7e3cd77d
fbshipit-source-id: bfaa5203fc7d998635c3c3f398cc180c7e3cd77d