Fix for the inliner
Summary:
The inliner needs to update the field or method ref in an instruction when the def behind the ref is modified (visibility/accessability).
Consider the following:
class A { void m() {} }
class B extends A {}
and a function
void f() {
..
invoke B.m()
..
}
if f() gets inlined the call to B.m() is checked and resolved to A.m(), then A.m() is made public so it can be used from the context it was moved to.
However the invoke is still using B.m() with B being package protected and the verifier chokes.
So we need to change the instruction to use to right ref

Reviewed By: newobj

Differential Revision: D3249667

fb-gh-sync-id: a35951d981dcbfcb99b65949e83fc5b3d92f161b
fbshipit-source-id: a35951d981dcbfcb99b65949e83fc5b3d92f161b