Refactor configuration handling
Summary:
Create a PassConfig abstraction that holds the config data for each
pass and can yield well-typed values on request.  This eliminates a bunch of
duplicated code for retrieving values from folly::dynamic in each pass, makes it
obvious what options exist for each pass, and reduces the surface area of
folly::dynamic.

Reviewed By: newobj

Differential Revision: D3235894

fb-gh-sync-id: 1094575fb561f63b8aa8eabacbff67686290350a
fbshipit-source-id: 1094575fb561f63b8aa8eabacbff67686290350a