Skip over relocation candidate targets that have no class data
Summary:
At output emission time we check if a class has got class data and if it does not we do not emit any class data item (methods, fields, etc.), which effectivley makes the class an impossible candidate for any kind of relocation.
This fix is local to StaticReolo and a more proper fix should be to allow emitting classes with no class data, or, more precisely, to set the class data flag and emit things correctly. However I have not investigated what it takes to do that so for now I am just killing the target class in StaticRelo if the target class had no class data.

build[automation_fbandroid_ics_redex,automation_fbandroid_art_redex,automation_fbandroid_redex]

Reviewed By: newobj

Differential Revision: D3303808

fbshipit-source-id: 03d5145dd3abbcc780b1b41503163cfaba06ff48