Parse and ignore -dontnote
Summary:
It appears that the Android SDK has updated their proguard and there are now
rules we are unable to parse in v25.1.6 of the Android Tools package.

This diff parses `-dontnote` directives. Sadly, this is still insufficient to
parse the whole official proguard-android.txt (P56375910) but it moves us past
this failure and onto the failure to parse the new `-keepclassmembers` with `%`.

Reviewed By: bertmaher

Differential Revision: D3336763

fbshipit-source-id: 076fbac0c235239846d7f051d5e02698988b1e02