Stop using proguard rules for deletion filter
Summary:
It looks like we weren't actually using these prior to D3251091.  That
diff removed the clearing of referenced state following the proguard rule
parsing, which was sort of accidental, but it's not clear why we were clearing
them in the first place.

build[automation_fbandroid_ics_,automation_fbandroid_art_]

Reviewed By: onomatopoeia

Differential Revision: D3337641

fbshipit-source-id: cea487fb1507483696fd15123f09017f6f00c900