Write out deobfuscated method names to the method ids file
Summary:
The current method IDs mapping file uses the final method name in the dex. This
is the obfuscated post-proguard name, so it's not entirely useful.

This diff extends ProguardMap with the inverse map (obfuscated -> deobfuscated
name) and pipes it through to DexOutput, where it is in turn used to write out
the deobfuscated names in the mapping file.

Also, this diff changes a few methods' constness to allow `const ProguardMap&`
to be useful.

Reviewed By: bertmaher

Differential Revision: D3338713

fbshipit-source-id: 0b2d49997a97727b0736f3535dd981321c469d74