FinalInline local fix
Summary:
The code here is effectively a revert of the previous fix and make a localized fix to FinalInline.
Essentially we allow removal of fields for every class that starts with Lcom/facebook/R$ which account for almost all the missing removal.
We go from
Removable fields 57925/59182
to
Removable fields 57890/59084

What concerns me of the original fix is that it is more intrusive by changing a blacklist that we run with in alpha and beta. And we run with it on all applications (messenger, pma, etc.), so I am not sure avout the ramifications.
This fix is increasing the number of fields removed but it is actually removing a subset of the original final inline which seems safe.

Overall this is not the real diff because it does not apply cleanly on beta so if everything works I am going to revert the original fix and make a new diff for this.

build[automation_fbandroid_ics_,automation_fbandroid_art_,automation_fbandroid_redexunstable,automation_messenger_ics_redex,automation_messenger_art_redex,automation_messenger_ics_xhdpi_armv7_release,automation_messenger_art_xhdpi_armv7_release]

Reviewed By: onomatopoeia

Differential Revision: D3358547

fbshipit-source-id: 69ef9451bfd157381aac90573b70277a13ef5ef8