Remove FinalInline annotations and class members.
Summary:
The FinalInline pass has some very specific set of annotations and class members in the configuration.
In the current setup, there is a global model using the keep_annos, which handles the class annotations and there is no way to handle class members which used by FinalInline pass.

The changeset moves them to a global place as opposed to optimization pass. I have added the required changes in ReachableClass method which parses the fb4a.config.

Reviewed By: newobj

Differential Revision: D3343781

fbshipit-source-id: 50cf47de60de8666bcfbc3dc653678e56f53b9fc