Fix the inliner when computing interface methods from derived classes
Summary:
When computing the set of interface methods coming from derived classes we were overriding the "existing ones with the same name" across multiple children.
The unit test InterfacesWithImplInBaseMultipleClassesAndOverloads3Final shows the problem.
When the fix is not there that unit tests fails at times (not always as order matters) because it marks the interface method in the base as non virtual and thus the inliner may inline it.
Task 11410434 describes the specific issues for fb4a.

There are a lot of comments in the code and I need them for a couple more days until I write all the unit tests that check all possible scenarios.
Also we need to fix RedexContext so we can create and destroy them across unit tests, at the moment we can only run a unit test at a time.

The reason why that is not completely broken is because method implementations for interfaces in base classes not implementing the interface are rare. On top of that when those are around they should be inlined which is not always the case.
Even more in some situations this depends on the order children are computed so it's not consistent.

I tried a run of fb4a with and without the fix and there is no change in the number of inlined functions which suggests that things are only problematic for the specific case when the build gets broken (unlikely right now). Crashes seem to imply this is not a problem except the specific instance in the task.

The change is pretty small but the implementation of the unit test and the fixes related to running the unit tests and the change applied to redex and redex-stable make it look big.

build[automation_fbandroid_ics_,automation_fbandroid_art_,automation_fbandroid_redexunstable,automation_messenger_ics_redex,automation_messenger_art_redex,automation_messenger_ics_xhdpi_armv7_release,automation_messenger_art_xhdpi_armv7_release]

Reviewed By: satnam6502

Differential Revision: D3364984

fbshipit-source-id: 248983627da001a7ac422ca6338293152c91d93b