Restrict AnnoKill to touch any class with certain annotations even if the annotation is in a member of the class
Summary:
AnnoKill is still removing too many annotations and it's getting tricky to understand what to keep or remove.
With this diff we are going to be a bit more restrictive and for a set of annotations defined in the config we are going to check the class and all members and if any has the annotation we leave the class alone.
That is mostly to cover the JSON serialization which is killing us at the moment.

build[automation_fbandroid_ics_,automation_fbandroid_art_,automation_fbandroid_redexunstable]

Reviewed By: newobj

Differential Revision: D3390803

fbshipit-source-id: 237934c15f9d63ccca3b2e33ebd3bd95a68bda8f