renamer: preserve a token '$' if original class might have been inner class
Summary:
Android instrumentation test runner class loads all classes in the APK while looking for tests to run. It purposely skips "inner classes" as indicated by a '$' in the class name. This is presumably due to the common pattern of putting API-gated stuff in an inner class.

Renamer blows away '$' when renaming inner classes, causing the test runner to classload classes it normally wouldn't. I was running into a problem where one such class extends ClassValue<T>, which is java7 only and not available in Android obviously.

In this approach, we include a token '$' when renaming something that appears to be a formerly inner class. This causes the instrumentation test runner to exclude it from classloading.

Reviewed By: dariorussi

Differential Revision: D3413077

fbshipit-source-id: 43307a5e61a3744148eebab89c0ebef40121600f