Fix bug in running ConstantPropagation on fb4a
Summary: This bug was due to missing constraint in the code. Some branch was propagated with constant value incorrectly. As a quick fix, I make the constant propagation condition as strict as possible, i.e. the if-eqz instruction is only propagated when it's right after a const/4 instruction. If there's any move instruction that changes the value of a dest register loaded by const/4, this register is marked as unavailable for propagation. This update will eliminate running error. But sadly, this will remove no conditional branch (because the the condition for propagation is too strict). I'll build on this version to add more conditions that can help improve the rate of code being removed.

Reviewed By: satnam6502

Differential Revision: D3408269

fbshipit-source-id: 974ae82b763f3221ff19f213d400f8ecc7c66af6