Retain old line number behavior if config does not ask for line mapping
Summary:
dariorussi mentioned that we should have a way to toggle the new behavior on and
off.

It also seems silly to remap the line numbers without emitting the file that
allows us to recover the originals, so I'm having redex emit the old line
numbers in that case.

Note that the positions will still be wrong for code that gets inlined from a
different file; the stack traces will instead show that the inlined code came
from the file that it got inserted into. Ideally we'd replace them by
DEX_NO_INDEX instead of giving misleading line numbers. However, this is an
existing bug, so I'm not going to the trouble to fix it just yet.

build[automation_fbandroid_ics_,automation_fbandroid_art_]

Reviewed By: dariorussi

Differential Revision: D3398204

fbshipit-source-id: f40d6f7ab40295f180df4bf529989ef052dc90fb