Dedup catch handlers
Summary:
We were duplicating catch handlers unnecessarily. E.g. suppose the original dex
input had two try items pointing to the same catch handler. Since our internal
representation of DexTryItem stores the catch handlers directly instead of as a
pointer, we would create two copies of catch handlers to represent this. Then
when we wrote out the redexed dex we would write out two identical catch handler
items.

This diff dedups the catch handlers before output to fix this.

It also changes our internal representation of catch handlers: instead of having
a separate field to represent the catch-all handler, we instead store it like any
other handler, but with null for the pointer to the exception type. This turns out
to be easier to manipulate by e.g. the inliner.

Reviewed By: dariorussi

Differential Revision: D3410957

fbshipit-source-id: e08e595f6aabec78c16c565fb79616bc2615feed