Remove the try-in-caller restriction for the inliner
Summary:
When the caller has a try and the callee has an if-else, some care is required.
Namely, consider the following pseudo-bytecode:

  [callee]
  ifeq
  ...
  return-void
  throw

  [caller]
  <TRY_START>
  invoke-direct callee
  <TRY_END>
  return-void

(TRY_START and TRY_END are not actually bytecodes, but MethodItemEntries that we
use to represent try blocks in our FatMethods). After inlining, we must
remember to add another try block:

  [callee]
  ifeq
  ...
  return-void
  throw

  [caller]
  <TRY_START>
  ifeq
  ...
  <TRY_END>
  return-void
  <TRY_START>
  throw
  <TRY_END>

This diff does that.

For ease of rollout, the restriction is still enabled by default; it can be enabled via a config setting.

Reviewed By: dariorussi

Differential Revision: D3410970

fbshipit-source-id: 6dfe157e6e9ad9383fe6004d689a2e4659c158dc