Rename methods when mutating their protos / classes
Summary:
While working on {D3456101} I discovered a bug in my implementation
which was caused by overwriting existing methods while mutating a different one
of the same name. I fixed that bug and added an assertion to check for this
case, but it turned out that other optimizations would trip on it -- probably
indicating some latent bugs. I've fixed it by making the other opts rename the
methods that they mutate (though since my knowledge of those opts is limited
I'm not 100% sure it's the best fix -- do let me know)

Reviewed By: dariorussi

Differential Revision: D3456197

fbshipit-source-id: cf8cc83554e95dc81b72f80c3095a3bffd9ee367