per-pass custom metrics
Summary:
Enables passes to emit "custom" metrics that are not necessarily reflected in apk structure. For example, the renamer pass (as seen in this diff) can emit the number of classes it renamed.

A lot of optimizations make changes not directly reflected in dex item/ref counts, which are never the less still important to quantify. Now we can do that.

All passes now receive a PassManager, which has methods for manipulating metrics.

**NOTE** This will require another diff for stats bot to actually display the results.

Reviewed By: onomatopoeia

Differential Revision: D3546173

fbshipit-source-id: 058d9d47941989a9a21abb73e368b74c02e266b6