Avoid duplicating encoded_array data
Summary:
I ran our instr APK through redex with the `nopasses` config and observed that
the APK increased slightly in size. Further investigation pointed at two causes:

1. The locator strings
2. The fact that our encoded_arrays get duplicated unnecessarily.

This diff disables the locator strings in the nopasses config and fixes the
duplication. It also adds a test to ensure that running with `nopasses` won't
increase the dex size.

Note that this test would have revealed {D3410957} as well.

Reviewed By: bertmaher

Differential Revision: D3520538

fbshipit-source-id: 9be8f36162a425d9187479f15a320d761309a64d