add option to treat primary dex like other dexes in interdex pass
Summary:
For art builds the primary dex has no special status when it comes to first coldstart or optimization criteria. The system takes care of all dex unpacking and compiles all dexes together. So the restrictions and rules for the primary dex are not needed.
Here I add a flag which makes interdex treat all dexes the same. This results in the first dex being full sized and 1K clases (in fb4a) which are not in the coldstart list being moved from a dex in which they are going to be loaded during coldstart to a non-coldstart dex. File size savings are ~600KB for the coldstart dexes.

build[automation_fbandroid_redexunstable_ics_,automation_fbandroid_redexunstable_art_]

Reviewed By: bertmaher

Differential Revision: D3508070

fbshipit-source-id: d7faf127074bd779fe218781b8dd7af9aa8fcbf7