regalloc: Change how we calculate interference graph
Summary:
Previously, the register allocator would put an edge between two nodes
of the graph if they were simultaneously live. However, if some register R is
assigned to but that value never gets used, R should still interfere with all
other registers that are live-out at that instruction. (A dead code elimination
pass would remove such an instruction, of course, but this is nonetheless more
correct.)

Also change the interference graph to be stored as a vector of bitvectors, which
makes certain operations faster.

Reviewed By: bertmaher

Differential Revision: D3554709

fbshipit-source-id: 7bb99d500a09241727f8df6b06212ce43d150caf