make simple inline deterministic (brute force version)
Summary:
SimpleInline was causing non-deterministic redex buids. You would see this in stats bot with weird little +1, +2, -2, -1 changes that made no sense. I also verified this in a loop of redexing an apk and using compare_bits.

I tried to fix this "surgically" but couldn't seem to find the exact unordered thing causing the problem, so I took a hammer and made all the sets/maps in simple inliner ordered. After spending over a day on this, I need to move on :(

Reviewed By: grfrost, int3

Differential Revision: D3595104

fbshipit-source-id: a4ebcdf8a42fdffea317d867537dbee07b4517ab