Generate the LivenessMap in a separate pass
Summary:
Previously, we recorded the live-out data of each instruction through each
iteration of the dataflow analysis, but we only care about the live-out data
from the last iteration (when we know we have reached the fixed point). That
means we were allocating lots of dynamic_bitsets just to free them a short
while later, and it turns out that it is extremely expensive.

Reviewed By: bertmaher

Differential Revision: D3592844

fbshipit-source-id: aa22362e61691e11fe47f44220879576e8e4deb9