Use pointer comparison for m_make_static
Summary:
{D3595104} made the inliner deterministic by replacing unordered sets (with
pointers for keys) with ordered sets (which hashed the method / field signatures
for keys, since memory addresses are not consistent across runs).

However, this led to a very subtle bug with DexMethod::mutate(). `mutate` would
modify the signature of a method in-place. To maintain correctness, a set
of methods ordered by signature must be rebalanced after `mutate` is called --
otherwise we might e.g. incorrectly report a method as 'not in the set'.

I've just done the easiest solution for now, which is to revert back to a
pointer comparison. I haven't yet tested whether this brings back
nondeterminism. (There's also the question of whether `mutate()` is a good API
to have.) But I think we can deal with that later.

Reviewed By: dariorussi

Differential Revision: D3611292

fbshipit-source-id: 5390618ff94e87460ab92f81deeafa0a556d3847