Reduce the number of try_syncs required
Summary:
try_sync has to iterate to a fixed point because it needs to adjust branch
opcodes to encode their new offsets, and the act of adjusting those opcodes can
change the offsets of other branch opcodes. Note that if a branch opcode needs
adjustment, it's probably not the only one:

  a:
  b:
  invoke-direct foo
  goto <a>
  nop
  goto <b>

Suppose the invoke-direct gets inlined and adds more than 2^7-1 in opcode size
to the caller. Then both gotos that follow it must be adjusted to goto/16. The
old implementation would do this adjustment in two iterations of try_sync; now
we take only one.

Reviewed By: bertmaher

Differential Revision: D3623810

fbshipit-source-id: 954c1841c2b672a19a7501182fe45f10e6c98112