Adds rxapk, a tool for working with apk files
Summary:
It's a pretty common task to want to do something with the dex files inside an
apk. Since there are several different ways of packing dex files into an apk,
this can be a pain. This diff provides an easy way to create lightweight tools
that extract the dexes from an apk and then do something with them. It also
adds simple tools for running dexdump, redexdump, and for printing information
about the total dex code size.

Let me know if you have a better name than "rxapk". I wanted it to be short and
unique and make it clear that the tool operates on APKs, but I'm not super
happy with the current name so better names for the bike shed are welcome.

Reviewed By: bertmaher

Differential Revision: D3636295

fbshipit-source-id: 01a5540d04d1f2de4235aa02b72403cfaa75b762