delsuper: Clean up all references to the old method
Summary:
The current implementation of delsuper removes the method from its containing
class, but doesn't change the instructions that reference it. If the pass is run
in isolation it works fine -- any `invoke-virtual` instruction will
automatically resolve to the super method at runtime. However, there are two
issues with this:

1. We now have a bunch of invoke-virtual instructions that reference a method
  that *used* to be concrete, have a DexCode, etc. Since DelSuper did not update
  the method to be non-concrete, SimpleInline (with the right configs and input
  code) would get confused and end up deleting methods it should not have.

2. It probably means that extra, unnecessary MethodRefs remain behind.

So to solve two problems at once I opted to remove all references to the
methods marked for deletion.

Reviewed By: dariorussi

Differential Revision: D3667550

fbshipit-source-id: 77f7bfc194d1049519e390416e4878cab70b0855