Fix `make install`
Summary:
My refactoring which moved code out of redex.py into separate files broke `make
install` because that command just copies redex.py, not the files it depends
on. With this fix, the python files it depends on will also be copied.

Note: this isn't a very elegant solution, but I think it's acceptable since the
majority of people aren't doing "make install" but rather are just running
redex.py directly from the redex directory.

Reviewed By: bertmaher

Differential Revision: D3657813

fbshipit-source-id: 016cd07b7823fcfbf8b1774faff462ce9e65233f