Don't clean up temp directories if redex-all fails
Summary:
It's a pain to have to run redex again with -d to get a command line
for gdb.  Just don't clean up the temp directory in this case.  Work around
atexit by raising a signal, which isn't the most beautiful thing we could do,
but gets the job done.

Reviewed By: int3

Differential Revision: D3690745

fbshipit-source-id: 01bae0e79e2e13d10eb60f06303f6c644eddbe56