Deobfuscate names more reliably
Summary:
Our method-ids map had a lot of unobfuscated names in it, for a couple
reasons:

1. Optimizations that change the names of classes or signature of methods break
the string-based lookup used by ProguardMap.  We can solve this by caching the
deobfuscated name so it gets carried around automatically.  In a few places we
create new methods using old ones as a template, and there we have to
explicitly carry the deobf name forward.  (That part could probably be
cleaner.)

2. Looking up field refs without a corresponding def fails, because Proguard
only gives you a mapping for defs.  To work around that, we resolve the method
to a def so we can find its cached deobf name.

Reviewed By: newobj

Differential Revision: D3686353

fbshipit-source-id: c7bdaa4f67db4a60c587b5d40081eb4634cca60c