Reverted commit D3686353
Summary:
Our method-ids map had a lot of unobfuscated names in it, for a couple
reasons:

1. Optimizations that change the names of classes or signature of methods break
the string-based lookup used by ProguardMap.  We can solve this by caching the
deobfuscated name so it gets carried around automatically.  In a few places we
create new methods using old ones as a template, and there we have to
explicitly carry the deobf name forward.  (That part could probably be
cleaner.)

2. Looking up field refs without a corresponding def fails, because Proguard
only gives you a mapping for defs.  To work around that, we resolve the method
to a def so we can find its cached deobf name.

Reviewed By: newobj

Differential Revision: D3686353

fbshipit-source-id: 82745bb66bcdfa3913770c8758136127f9a7b8e2