Bugfix of constants in InterDex
Summary:
We were calling sizeof(thing) to get the length of a string declared
as a pointer, which always gives us the pointer size.  If we declare it as an
array this works properly.  This doesn't end up mattering in practice because
there are no classes that start with "Lseconda", but still.

Reviewed By: onomatopoeia

Differential Revision: D3758666

fbshipit-source-id: f19559fe2f6fa862b0557734fcc5d7e3c94da0c7