Copy rstate when "cloning" methods/fields
Summary:
Summary When we're cloning methods in singleimpl we need to copy all their
random properties, including rstate (and I missed one deobfuscated_name
earlier).  We should probably have a helper method for this but I don't really
understand the cloning process in singleimpl so I prefer this incremental
change.

Reviewed By: dariorussi

Differential Revision: D3778985

fbshipit-source-id: 4cfcf82af392f86b5f047c2763e595805ee3fd30