Enable dual parsing of ProGuard configs
Summary:
This diff enables dual parsing of ProGuard configs using both the
current ProGuard parser (which is the default) and the new ProGuard
parser (which is used to compute keep information, but this keep
information is not used in optimizations yet, until the parser has
run for a while without issue). Gradually the ProGuard parsing
primary tasks will be taken over by the new parser, and the old
parser will be deleted.

Reviewed By: bertmaher

Differential Revision: D3829917

fbshipit-source-id: 3a7b187c071a7424ae59889ab32587ea6a7670f0