Remove unreachable fields/methods/classes
Summary:
This pass uses a mark-sweep algorithm to remove unreachable parts of
the dexes.  It starts with the Proguard seeds (eventually it will use the -keep
rules), and does a pass to mark everything reachable from those.  Everything
not marked is removed by the sweep pass.

The most notable aspect of the marking pass is that we mark any
override/implementation of a marked method.  This rule means we need a reverse
inheritance graph to perform the marking pass.  Building that graph is by far
the most time-consuming part of this pass.

Note that the pass currently has no representation of conditional keep rules
(-keepclassmembers and -keepclasseswithmembers), because the seeds don't
capture this information.  Seeds over-approximate this property, so to perform
more precise deletion, we will need to add conditional visitation of
classes/members.

Reviewed By: dariorussi

Differential Revision: D3638929

fbshipit-source-id: 836018fd429b0be832e822627bd966aaf4f576af