Add instr test to check that shrinking is actually doing something
Summary:
I wanted to port all the integration tests over to this framework, but
I got cold feet since it's tricky to make sure that the proguard rules produce
the seeds you wanted.  But this seemed worthwhile to land anyways (especially
because the Test annotation stuff took me a long time to figure out).

Reviewed By: satnam6502

Differential Revision: D3845489

fbshipit-source-id: 60504ea5dd7f63f0ae6584485baa653bb4c53bac