Rewrite opcode matcher
Summary:
I rewrote the bytecode matcher part of the Matcher library in redex, to address the following:

**Simpler**
The old system had this crazy convoluted idea of DexOpcodeSeq and being able to match a single instruction versus a collection of instructions. It was weird. It's all been simplified now.

**Visitable**
The old system just matched "yes/no", it had no way to visit the sequence that fully matched the pattern. This was necessary for use cases that will come up in forthcoming diffs, like "find all instances Class.forName of a string constant" Observe this in D3582899:

    match_opcodes(scope, match, [&](const DexMethod* meth, size_t n, DexInstruction** insns){
      DexOpcodeString* const_string = (DexOpcodeString*)insns[0];
      DexOpcodeMethod* invoke_static = (DexOpcodeMethod*)insns[1];
      // Make sure that the registers agree
      if (const_string->dest() == invoke_static->src(0)) {
        TRACE(PGR, 1, "Class.forName: %s\n", const_string->get_string()->c_str());
      }
    });

After we match our given bytecode pattern, we receive the opcodes and are able to ensure that Class.forName is actually using the register that the const-string was loaded into!

**Consistent w/ Other Refineable Predicates**
The old system didn't allow any further predicates inside the DexInstruction matchers. It was just kind of a dead end. The "langauge" is now capable of taking refining predicates that further refined instructions beyond just the opcode.  Again, from D3582899:

    auto match = std::make_tuple(
      m::const_string(/* const-string {vX}, <any string> */),
      m::invoke_static(/* invoke-static {vX}, java.lang.Class;.forName */
        m::opcode_method(m::named<DexMethod>("forName") && m::on_class<DexMethod>("Ljava/lang/Class;"))
        && m::has_n_args(1))
    );

We are able to refine the invoke-static matcher to specifically be an invoke-static of Class.forName with 1 arg.

Reviewed By: bertmaher

Differential Revision: D3582816

fbshipit-source-id: 9a222c99ac2b053d9427c0061afcd073cf94060a