Implement extends and implements ProGuard clause
Summary: This diff implements support for the extends and implements clauses for keep rules. It also implements filtering by annotation type for the extended or implemented type. The diff includes test cases which check for classes in dependent jars (a dependent jar is built as part of the test) and checks for chains that alternate between extends and implements.

Reviewed By: bertmaher

Differential Revision: D3845790

fbshipit-source-id: 3bc25786887669744bd052663b593a93f17cb02e