Change how may-throw opcodes are handled in CFG (fixes LocalDCE)
Summary:
We insert one MFLOW_FALLTHROUGH before every MFLOW_OPCODE that
could potentially throw, and set the throwing_mie field to point to that
opcode. The MFLOW_FALLTHROUGH will then be at the end of its basic block
and the MFLOW_OPCODE will be at the start of the next one. build_cfg
will treat the MFLOW_FALLTHROUGH as potentially throwing and add edges
from its BB to the BB of any catch handler, but it will treat the
MFLOW_OPCODE is non-throwing. This is desirable for dataflow analysis since
we do not want to e.g. consider a register to be defined if the opcode
that is supposed to define it ends up throwing an exception. E.g. suppose
we have the opcodes

  const v0, 123
  new-array v0, v1

If new-array throws an OutOfMemoryError and control flow jumps to some
handler within the same method, then v0 will still contain the value 123
instead of an array reference. So we want the control flow edge to be
placed before the new-array instruction. Placing that edge right at the
const instruction would be strange -- `const` doesn't throw -- so we
insert the MFLOW_FALLTHROUGH entry to make it clearer.

Previously, we were inserting the edge to the handler at the new-array
instruction, which caused LocalDCE to wrongly conclude that `const` was a dead
instruction.

One catch to all of this is that SimpleInline's use_liveness option currently
relies on the old behavior, and it seemed nontrivial to change that. Also it
seems like the effort would be better spent on getting the register allocator to
work. As such I've preserved the old behavior under a flag.

Reviewed By: dariorussi

Differential Revision: D3903754

fbshipit-source-id: b010ee49a249ef08624ac583051beeb726205b99